requirements.txt
pygame~=2.0.2
setuptools~=58.0.4
PyYAML~=5.4.1
PyTMX~=3.27
pygame-gui~=0.5.7
pyioc3~=1.3.0
appdirs~=1.4.4
